college functionality done

+++ src/finalproject/SchoolYear.java
@@ -9,7 +9,7 @@ import javax.swing.table.DefaultTableModel;
 import javax.swing.table.TableModel;
 import finalproject.Table.TableActionCellRender;
 import javax.swing.text.AbstractDocument;
-import finalproject.FieldValidation.FourDigitFilter;
+import finalproject.FieldValidation.IntDigitFilter;
 
 public class SchoolYear extends javax.swing.JFrame {
     Connection conn = null; 
@@ -34,7 +34,7 @@ public class SchoolYear extends javax.swing.JFrame {
     
    private void applyDocumentFilter(JTextField textField, boolean apply) {
         if (apply) {
-            ((AbstractDocument) textField.getDocument()).setDocumentFilter(new FourDigitFilter());
+            ((AbstractDocument) textField.getDocument()).setDocumentFilter(new IntDigitFilter(4));
         } else {
             ((AbstractDocument) textField.getDocument()).setDocumentFilter(null);
         }
@@ -255,6 +255,10 @@ public class SchoolYear extends javax.swing.JFrame {
     }// </editor-fold>//GEN-END:initComponents
 
     private void btnAddMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnAddMouseClicked
+        if (txtStart.getText().isEmpty()||txtEnd.getText().isEmpty()) {
+            JOptionPane.showMessageDialog(null, "Please fill in all fields.");
+            return;
+        }
         String query = "INSERT INTO schoolyear (syear) VALUES (?)"; 
         try (Connection conn = ConnectPLMDB.Connect(); 
              PreparedStatement pstmt = conn.prepareStatement(query)) { 
@@ -304,10 +308,10 @@ public class SchoolYear extends javax.swing.JFrame {
 
     private void btnDeleteMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnDeleteMouseClicked
         // TODO add your handling code here:
-    }//GEN-LAST:event_btnDeleteMouseClicked
-
-    private void btnDeleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnDeleteActionPerformed
-        // TODO add your handling code here:
+        if (txtStart.getText().isEmpty()||txtEnd.getText().isEmpty()) {
+            JOptionPane.showMessageDialog(null, "Please enter the School Year to delete.");
+            return;
+        }
         String query = "DELETE FROM schoolyear "
                      + "WHERE syear = (?)"; 
         try (Connection conn = ConnectPLMDB.Connect(); 
@@ -320,6 +324,11 @@ public class SchoolYear extends javax.swing.JFrame {
                 JOptionPane.showMessageDialog(null, "Delete Existing Related Records First!"); 
         }
         refresh();
+    }//GEN-LAST:event_btnDeleteMouseClicked
+
+    private void btnDeleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnDeleteActionPerformed
+        // TODO add your handling code here:
+        
     }//GEN-LAST:event_btnDeleteActionPerformed
 
     public static void main(String args[]) {
